Move Separate By to Additional Tools
Move Separate BY to Additional tools, makes no sense having these smaller functions in it's own file.
